Refactored GameCompleteModal
start of day complete modal, which currently does not display
