+++ frontend/src/components/Modal/GameCompleteModal.jsx
@@ -1,10 +1,10 @@
-import "./GameCompleteModal.css"
+import "./Modal.css"
 import ReactDom from "react-dom"
 
 export default function GameCompleteModal({ score }) {
     return ReactDom.createPortal(<div className="modal-overlay">
         <div className="modal-content">
-            <div className="text">🎉Great job!🎉</div>
+            <div className="text">🎉Great Job!🎉</div>
             <div className="score-earned">${score}</div>
             <button onClick={() => window.location.href = "/"} className="home-button">
                 🏠 Back to home
